python-backend/app/requirements.txt
# FastAPI and Web Server
fastapi>=0.68.0
uvicorn>=0.15.0

# Database ORM and Migrations
sqlalchemy>=1.4.23
alembic>=1.7.1
psycopg2-binary>=2.9.1
pymongo>=4.6.1

# Data Validation and Settings Management
pydantic>=1.8.2
pydantic-settings>=2.0.0

# Environment Variables
python-dotenv>=0.19.0

# Authentication and Security
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.2

# File Handling and Processing
python-multipart>=0.0.5
pandas>=2.2.0
openpyxl>=3.1.2
xlrd>=2.0.1

# Data Validation
email-validator>=2.1.0
python-dateutil>=2.8.2

# Face Recognition and Image Processing
opencv-python>=4.8.0
numpy>=1.24.0
face-recognition>=1.3.0
dlib>=19.24.0
cloudinary>=1.33.0

# Machine Learning and Deep Learning
tensorflow>=2.13.0  # For advanced face recognition models
scikit-learn>=1.3.0  # For machine learning utilities

# # API Documentation
# python-markdown>=3.4.0
# mdx-math>=1.0.0

# Performance and Optimization
ujson>=5.8.0  # Faster JSON processing
orjson>=3.9.0  # High-performance JSON library
python-jose[cryptography]>=3.3.0  # For JWT handling with better crypto

# Testing and Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0  # For async HTTP requests in tests